-Agent improvement and added agent
-Added two new observations for agent (current platform ID and last platform ID):
They are used to inform the agent which platform he is currently on and which last platform he reached (it used because agent doesn't know his position in environment).
-Reward improvement:
The agent now will receive negative rewards if it returns to previous platforms that it has already passed (it's increasing arithmetically as returns (e.g. first previous platform -0.1, second -0.2)), is used to prevent the policy that routes agent back.
-Added agent model for medium levels with wall (500k steps)
Warning: Previous model won't work at this commit (use previous commits to test the models)